removing non-paylist files upon finishing download

+++ firmware/main/playlist/track.h
@@ -6,7 +6,7 @@
 
 typedef struct {
   const char *filename;
-  uint32_t id;
+  int64_t id;
 } Track_t;
 
 #endif
